Regions may progress only when they get larger
This prevents unnecessary movement of sticky blocks in cases when there
is a backlog of regions smaller or equal to the current region. I don't
think there is a case where a smaller region could fix a situation where
there is a lack of space.

Since this unnecessary movement was previously already prevented for
orphan and widow prevention, this commit also restores that behavior.